feat: reorganize scheduling logic by moving generateEventsForDay and adding findNextStream functionality
